pyproject.toml
[project]
name = "virtualmoments"
version = "0.1.0"
description = "Package for building Astro page with screenshots from Steam"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argparse>=1.4.0",
    "beautifulsoup4>=4.14.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "selenium>=4.41.0",
]

[tool.setuptools.packages.find]
include = ["VirtualMoments*"]
